+++ src/framework/components/BaseDynamicComponent.ts
@@ -3,15 +3,11 @@ import type { DisplayItem } from "../../ui/homepage/data/types/DisplayItem.ts";
 import {
   clearSubscribers,
   createComponentStore,
-  getComponentStore, updateComponentStore,
+  updateComponentStore,
 } from "../state/data/ComponentStore.ts";
 import {
   initRequestStoresOnLoad,
 } from "../state/data/RequestStore.ts";
-import { EventHandlerAction } from "../state/update/event/EventHandlerAction.ts";
-import { BaseDispatcher } from "../state/update/BaseDispatcher.ts";
-import { EventThunk } from "../state/update/event/EventThunk.ts";
-import type { EventHandlerThunkConfig } from "../state/update/event/types/EventHandlerThunkConfig.ts";
 import {
   type ComponentLoadConfig,
   type ThunkDispatcherConfig,
@@ -22,23 +18,17 @@ import {
   getGlobalStateValue,
   subscribeComponentToGlobalField,
 } from "../state/data/GlobalStore.ts";
-import type { EventValidationResult } from "../state/update/event/types/EventValidationResult.ts";
 import type {FormItemConfig} from "./types/FormItemConfig.ts";
 import  {FormSelector} from "../FormSelector.ts";
-
-type EventConfig = {
-  eventType: string;
-  eventFunction: (e: Event) => any;
-};
+import {EventHandlerData} from "../handler/EventHandlerData.ts";
+import {generateLoadingIndicator} from "./utils/StatusIndicators.ts";
 
 export abstract class BaseDynamicComponent extends HTMLElement {
-  componentStoreName: string;
-  eventHandlerConfig: Record<string, EventConfig>;
-  eventTagIdCount = 0;
 
+  eventHandlerData:EventHandlerData;
+
+  componentStoreName: string
   requestStoreReducer?: BaseThunk;
-
-  instanceId: number;
 
   dependenciesLoaded: boolean = true;
   componentLoadConfig: ComponentLoadConfig | undefined = undefined;
@@ -49,13 +39,13 @@ export abstract class BaseDynamicComponent extends HTMLElement {
   constructor(componentStoreName: string, loadConfig?: ComponentLoadConfig) {
     super();
 
-    this.instanceId = BaseDynamicComponent.instanceCount;
     BaseDynamicComponent.instanceCount++;
-    this.eventHandlerConfig = {};
     this.componentStoreName = `${componentStoreName}-${BaseDynamicComponent.instanceCount}`;
 
     this.formSelector = new FormSelector();
     createComponentStore(this.componentStoreName, this);
+
+    this.eventHandlerData = new EventHandlerData(`data-${this.componentStoreName}-element-id`);
 
     if (loadConfig) {
       const self = this;
@@ -79,7 +69,6 @@ export abstract class BaseDynamicComponent extends HTMLElement {
         ) {
           subscribeComponentToGlobalField(self, fieldName);
           self.componentLoadConfig = loadConfig;
-
         });
       } else {
         initRequestStoresOnLoad(loadConfig);
@@ -107,78 +96,27 @@ export abstract class BaseDynamicComponent extends HTMLElement {
   }
 
   updateWithStoreData(data: any) {
-    this.eventHandlerConfig = {};
-    this.eventTagIdCount = 0;
-
+    this.eventHandlerData.resetData();
     this.generateAndSaveHTML(data);
-    this.attachEventHandlersToDom();
+
+    if(this.shadowRoot){
+      this.formSelector.setShadowRoot(this.shadowRoot);
+    }
+
+    this.eventHandlerData.attachEventHandlersToDom(this.shadowRoot);
   }
 
   updateStore(data: any) {
     updateComponentStore(this.componentStoreName, function(){return data})
   }
 
-  attachEventHandlersToDom() {
-    const eventHandlers = this.eventHandlerConfig;
-    const elementIdTag = this.getElementIdTag();
-
-    const addEventHandler = function (item: Element) {
-      const id = item.getAttribute(elementIdTag) ?? "";
-      const eventConfig = eventHandlers[id];
-      item.addEventListener(eventConfig.eventType, eventConfig.eventFunction);
-    };
-
-    if (this.shadowRoot) {
-      this.formSelector.setShadowRoot(this.shadowRoot);
-      this.shadowRoot?.querySelectorAll(`[${elementIdTag}]`).forEach(function (
-        item: Element,
-      ) {
-        addEventHandler(item);
-      });
-    } else {
-      document.querySelectorAll(`[${elementIdTag}]`).forEach(function (
-        item: Element,
-      ) {
-        addEventHandler(item);
-      });
-    }
-  }
-
   generateAndSaveHTML(data: any) {
     if (!this.dependenciesLoaded) {
-      this.innerHTML = this.getLoadingIndicator();
+      this.innerHTML = generateLoadingIndicator();
     } else {
       this.formSelector.clearFormSelectors();
       this.innerHTML = this.render(data);
     }
-  }
-
-  getLoadingIndicator(){
-    return `<h1>Loading</h1>`
-  }
-
-  getElementIdTag() {
-    return `data-${this.componentStoreName}-element-id`;
-  }
-
-  generateErrorMessage(message: string | string[] | undefined){
-    if(Array.isArray(message)){
-      let html = ''
-      message.forEach((item)=>{
-        html+=`<p class="error-message">${item.trim()}</p>`
-
-      })
-      return html;
-    }
-    return `
-      <p class="error-message">${message? message.trim() : ""}</p>
-        `
-  }
-
-  generateSuccessMessage(message: string | undefined){
-    return `
-      ${message  ? `<p class="success-message">${message}</p>` : ''}
-    `
   }
 
   generateInputFormItem(formConfig:FormItemConfig){
@@ -189,112 +127,16 @@ export abstract class BaseDynamicComponent extends HTMLElement {
     return this.formSelector.generateTextInputFormItem(formConfig);
   }
 
-  saveEventHandler(
-    eventFunction: (e: Event) => any,
-    eventType: string,
-  ): string {
-    let id = `${this.getElementIdTag()}=${this.eventTagIdCount}`;
-
-    this.eventHandlerConfig[this.eventTagIdCount] = {
-      eventType: eventType,
-      eventFunction: eventFunction,
-    };
-    this.eventTagIdCount++;
-    return id;
-  }
-
   createOnChangeEvent(eventConfig: any) {
-    const eventHandler = BaseDynamicComponent.createHandler(
-      eventConfig,
-      this.formSelector,
-      this?.componentStoreName,
-    );
-    return this.saveEventHandler(eventHandler, "change");
+    return this.eventHandlerData.createOnChangeEvent(eventConfig, this.formSelector, this.componentStoreName);
   }
 
   createSubmitEvent(eventConfig: any) {
-    let eventHandler;
-    eventHandler = BaseDynamicComponent.createHandler(
-      eventConfig,
-      this.formSelector,
-      this?.componentStoreName,
-    );
-    return this.saveEventHandler(eventHandler, "submit");
+    return this.eventHandlerData.createSubmitEvent(eventConfig, this.formSelector, this.componentStoreName);
   }
 
   createClickEvent(eventConfig: any, params?: any) {
-    let eventHandler;
-    eventHandler = BaseDynamicComponent.createHandler(
-      eventConfig,
-      this.formSelector,
-      this.componentStoreName,
-      params,
-    );
-    return this.saveEventHandler(eventHandler, "click");
-  }
-
-  static createHandler(
-    eventConfig: EventHandlerThunkConfig,
-    formSelector: FormSelector,
-    componentStoreName?: string,
-    params?: any
-  ) {
-
-    const eventThunk = eventConfig.apiRequestThunk;
-    const storeToUpdate =
-      eventThunk?.getRequestStoreId();
-
-
-    const dispatchers: BaseDispatcher[] = [];
-    let componentStoreUpdate: any;
-
-
-    if (componentStoreName) {
-      componentStoreUpdate = new BaseDispatcher(
-        componentStoreName,
-        eventConfig.componentReducer,
-      );
-      dispatchers.push(componentStoreUpdate);
-    }
-
-
-    const handler = function (e: Event) {
-
-      e.preventDefault();
-
-      const request: EventHandlerAction = new EventHandlerAction(
-        eventConfig.eventHandler,
-        componentStoreName,
-        formSelector,
-        params
-      );
-
-      if(storeToUpdate){
-        const storeUpdate = new BaseDispatcher(storeToUpdate, (a: any): any => {
-          return a;
-        });
-        dispatchers.push(storeUpdate);
-      }
-
-      const eventUpdater: EventThunk = new EventThunk(request, dispatchers);
-
-      if (eventConfig.validator) {
-        const eventConfigValidator = eventConfig.validator;
-        const validator = function (): EventValidationResult {
-          const componentData = getComponentStore(componentStoreName ?? "");
-          return eventConfigValidator(formSelector, componentData);
-        };
-
-        eventUpdater.processEvent(e, validator).then((result: any) => {
-          if (result?.errorMessage) {
-            componentStoreUpdate.updateStore(result);
-          }
-        });
-      } else {
-        eventUpdater.processEvent(e);
-      }
-    };
-    return handler;
+    return this.eventHandlerData.createClickEvent(eventConfig, this.formSelector, this.componentStoreName, params);
   }
 
   disconnectedCallback(){
